update travis.yml & test script

+++ go.test.sh
@@ -3,10 +3,35 @@
 set -e
 echo "" > coverage.txt
 
-for d in $(go list ./... | grep -v vendor | grep -v http2 | grep -v example); do
-    go test -v -race -coverprofile=profile.out -covermode=atomic $d
-    if [ -f profile.out ]; then
-        cat profile.out >> coverage.txt
-        rm profile.out
-    fi
-done
+if [[ $TRAVIS_GO_VERSION == 1.6* ]]
+then
+
+  for d in $(go list ./... | grep -v vendor | grep -v http2 | grep -v example); do
+      go test -v -race $d
+  done
+
+elif [[ $TRAVIS_GO_VERSION == 1.7* ]]
+then
+
+  for d in $(go list ./... | grep -v vendor | grep -v http2 | grep -v example); do
+      go test -v -race $d
+  done
+
+elif [[ $TRAVIS_GO_VERSION == 1.8* ]]
+then
+
+  for d in $(go list ./... | grep -v vendor | grep -v example); do
+      go test -v -race -coverprofile=profile.out -covermode=atomic $d
+      if [ -f profile.out ]; then
+          cat profile.out >> coverage.txt
+          rm profile.out
+      fi
+  done
+
+else
+
+  for d in $(go list ./... | grep -v vendor | grep -v example); do
+      go test -v -race $d
+  done
+
+fi
